requirements.txt
python==3.10
django==5.1
djangorestframework=3.15